+++ src/Zeta.FastEndpoints/ZetaGlobalPreProcessor.cs
@@ -1,5 +1,4 @@
 using FastEndpoints;
-using FluentValidation.Results;
 
 namespace Zeta.FastEndpoints;
 
@@ -21,9 +20,7 @@ internal sealed class ZetaGlobalPreProcessor<TRequest> : IGlobalPreProcessor
         var result = await _schema.ValidateAsync(request, validationContext);
         if (result.IsSuccess) return;
 
-        foreach (var error in result.Errors)
-            ctx.ValidationFailures.Add(
-                new ValidationFailure(error.PathString, error.Message) { ErrorCode = error.Code });
+        ctx.ValidationFailures.AddRange(result.Errors.ToValidationFailures());
 
         await ctx.HttpContext.Response.SendErrorsAsync(ctx.ValidationFailures, cancellation: ct);
     }